Move around and set up for mk
